moved CommandLineApplicationFacts to a dedicated file
